process-json/requirements.txt
google-cloud-storage==2.10.0
google-cloud-firestore
